fix: generators emit wrong numeric type for floating-point ranges (#78)
cddl2java, cddl2py, cddl2swift and cddl2kotlin all rendered a range like
(0.0..1.0) using the integer type, indistinguishable from (0..1), because
whole-valued float bounds collapse to integers under Number.isInteger.
Parser PR #77 added an IsFloat marker to numeric literal nodes to make the
distinction possible again; this wires that marker into all four
generators' range handling.

Adds a shared isFloatRange(range) helper to the cddl package rather than
duplicating the check per generator. It also handles Min/Max as a plain
number (the Range type's declared shape) in addition to the literal-node
object the parser actually emits, since at least one existing test double
(cddl2py's rangeRef()) already relied on the plain-number shape.

Along the way, fixed two related bugs in the same code paths beyond what
the issue's example covered:
- cddl2java always emitted Float for any range with an operator (e.g. a
  default value), even a genuinely integer range like `twist: (0..359)
  .default 0`-style declarations - now correctly Integer.
- cddl2py/cddl2swift/cddl2kotlin's generateVariable() had its own,
  separate hardcoded-int/Int/Long branch for a bare `MyRange = (0..1)`
  style variable, not just the resolveType() branch the issue quoted.

Verified against the real webdriver CDDL spec: pressure, tangentialPressure,
altitudeAngle, azimuthAngle and quality (all genuinely 0.0..1.0-style
ranges) now correctly render as Float/float/Double, while twist (0..359,
an integer range that happened to carry a default operator) now correctly
renders as Integer instead of the previous always-Float bug.

cddl2ts needs no change - TypeScript has no separate int/float numeric
type, so `number` was already correct.

+++ packages/cddl2java/tests/__fixtures__/BrowsingContext/ImageFormat.java
@@ -23,7 +23,7 @@ public class ImageFormat {
     /**
      * Creates a new ImageFormat instance
      */
-    public ImageFormat(String type, Integer quality) {
+    public ImageFormat(String type, Float quality) {
         this.type = type;
         this.quality = quality;
     }
@@ -38,13 +38,13 @@ public class ImageFormat {
         return this.type;
     }
 
-    private final Integer quality;
+    private final Float quality;
 
     /**
      * Gets the quality property
-     * @return Integer value
+     * @return Float value
      */
-    public Integer getQuality() {
+    public Float getQuality() {
         return this.quality;
     }
